% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpbnb.tmb-package.R
\docType{package}
\name{rpbnb.tmb-package}
\alias{rpbnb.tmb}
\alias{rpbnb.tmb-package}
\title{rpbnb.tmb: Random-Parameter Bivariate Negative Binomial with TMB}
\description{
Maximum-simulated-likelihood estimation of bivariate random-parameter
negative binomial models using Template Model Builder (TMB). Supports
Famoye/Sarmanov and discrete-copula (Frank, Gaussian, Clayton) dependence.
}
\author{
\strong{Maintainer}: Zhenyu Wang \email{wonstran@hotmail.com}

Authors:
\itemize{
  \item Zhenyu Wang \email{wonstran@hotmail.com}
}

}
\keyword{internal}
